app/requirements.txt
Django
psycopg2
djangorestframework
docker-py
django-cors-headers
